+++ esrally/driver/runner.py
@@ -114,6 +114,8 @@ class BulkIndex(Runner):
 
         * ``index``: name of the affected index. May be `None` if it could not be derived.
         * ``bulk-size``: bulk size, e.g. 5.000.
+        * ``bulk-request-size-bytes``: size of the full bulk requset in bytes
+        * ``total-document-size-bytes``: size of all documents contained in the bulk request in bytes
         * ``weight``: operation-agnostic representation of the bulk size (used internally by Rally for throughput calculation).
         * ``unit``: The unit in which to interpret ``bulk-size`` and ``weight``. Always "docs".
         * ``success``: A boolean indicating whether the bulk request has succeeded.
@@ -128,7 +130,8 @@ class BulkIndex(Runner):
         * ``shards_histogram``: An array of hashes where each hash has two keys: ``item-count`` contains the number of items to which a shard
           distribution applies and ``shards`` contains another hash with the actual distribution of ``total``, ``successful`` and ``failed``
           shards (see examples below).
-
+        * ``bulk-request-size-bytes``: Total size of the bulk request body in bytes.
+        * ``total-document-size-bytes``: Total size of all documents within the bulk request body in bytes.
 
         Here are a few examples:
 
@@ -164,6 +167,8 @@ class BulkIndex(Runner):
                 "weight": 5000,
                 "unit": "docs",
                 "bulk-size": 5000,
+                "bulk-request-size-bytes": 2250000,
+                "total-document-size-bytes": 2000000,
                 "success": True,
                 "success-count": 5000,
                 "error-count": 0,
@@ -193,6 +198,8 @@ class BulkIndex(Runner):
                 "weight": 5000,
                 "unit": "docs",
                 "bulk-size": 5000,
+                "bulk-request-size-bytes": 2250000,
+                "total-document-size-bytes": 2000000,
                 "success": False,
                 "success-count": 4000,
                 "error-count": 1000,
@@ -251,21 +258,35 @@ class BulkIndex(Runner):
         else:
             response = es.bulk(body=params["body"], index=index, doc_type=params["type"], params=bulk_params)
 
-        stats = self.detailed_stats(bulk_size, response) if detailed_results else self.simple_stats(bulk_size, response)
+        stats = self.detailed_stats(params, bulk_size, response) if detailed_results else self.simple_stats(bulk_size, response)
 
         meta_data = {
             "index": str(index) if index else None,
             "weight": bulk_size,
             "unit": "docs",
-            "bulk-size": bulk_size,
+            "bulk-size": bulk_size
         }
         meta_data.update(stats)
         return meta_data
 
-    def detailed_stats(self, bulk_size, response):
+    def detailed_stats(self, params, bulk_size, response):
         ops = {}
         shards_histogram = OrderedDict()
         bulk_error_count = 0
+        bulk_request_size_bytes = 0
+        total_document_size_bytes = 0
+
+        for line_number, data in enumerate(params["body"]):
+
+            line_size = len(data.encode('utf-8'))
+            if params["action_metadata_present"]:
+                if line_number % 2 == 1:
+                    total_document_size_bytes += line_size
+            else:
+                total_document_size_bytes += line_size
+
+            bulk_request_size_bytes += line_size
+
         for idx, item in enumerate(response["items"]):
             # there is only one (top-level) item
             op, data = next(iter(item.items()))
@@ -291,7 +312,9 @@ class BulkIndex(Runner):
             "success-count": bulk_size - bulk_error_count,
             "error-count": bulk_error_count,
             "ops": ops,
-            "shards_histogram": list(shards_histogram.values())
+            "shards_histogram": list(shards_histogram.values()),
+            "bulk-request-size-bytes": bulk_request_size_bytes,
+            "total-document-size-bytes": total_document_size_bytes
         }
 
     def simple_stats(self, bulk_size, response):
@@ -380,6 +403,18 @@ class Query(Runner):
                pages we will terminate earlier.
     * `items_per_page`: Number of items to retrieve per page.
 
+    Returned meta data
+
+    The following meta data are always returned:
+
+    * ``weight``: operation-agnostic representation of the "weight" of an operation (used internally by Rally for throughput calculation).
+                  Always 1 for normal queries and the number of retrieved pages for scroll queries.
+    * ``unit``: The unit in which to interpret ``weight``. Always "ops".
+    * ``hits``: Total number of hits for this operation.
+
+    For scroll queries we also return:
+
+    * ``pages``: Total number of pages that have been retrieved.
     """
 
     def __init__(self):
@@ -393,10 +428,22 @@ class Query(Runner):
             return self.request_body_query(es, params)
 
     def request_body_query(self, es, params):
-        es.search(index=params["index"], doc_type=params["type"], request_cache=params["use_request_cache"], body=params["body"])
-        return 1, "ops"
+        request_params = params.get("request_params", {})
+        r = es.search(
+            index=params["index"],
+            doc_type=params["type"],
+            request_cache=params["use_request_cache"],
+            body=params["body"],
+            **request_params)
+        hits = r["hits"]["total"]
+        return {
+            "weight": 1,
+            "unit": "ops",
+            "hits": hits,
+        }
 
     def scroll_query(self, es, params):
+        request_params = params.get("request_params", {})
         hits = 0
         retrieved_pages = 0
         self.es = es
@@ -412,7 +459,9 @@ class Query(Runner):
                     sort="_doc",
                     scroll="10s",
                     size=params["items_per_page"],
-                    request_cache=params["use_request_cache"])
+                    request_cache=params["use_request_cache"],
+                    **request_params
+                )
                 # This should only happen if we concurrently create an index and start searching
                 self.scroll_id = r.get("_scroll_id", None)
             else: