+++ esrally/racecontrol.py
@@ -66,6 +66,7 @@ class Benchmark:
         self.docker = docker
         self.actor_system = None
         self.mechanic = None
+        self.ignore_unknown_return = self.cfg.opts("system", "ignore.unknown.return")
 
     def _load_track(self):
         return track.load_track(self.cfg)
@@ -131,8 +132,13 @@ class Benchmark:
             result = self.actor_system.ask(main_driver,
                                            driver.StartBenchmark(self.cfg, self.race.track, self.metrics_store.meta_info, lap))
         except KeyboardInterrupt:
-            result = self.actor_system.ask(main_driver, driver.BenchmarkCancelled())
             logger.info("User has cancelled the benchmark.")
+            self.actor_system.send(main_driver, driver.BenchmarkCancelled())
+            return False
+        finally:
+            logger.info("Race control has received a benchmark result message. Terminating main driver actor.")
+            import thespian.actors
+            self.actor_system.tell(main_driver, thespian.actors.ActorExitRequest())
 
         if isinstance(result, driver.BenchmarkComplete):
             logger.info("Benchmark is complete.")
@@ -149,6 +155,7 @@ class Benchmark:
             logger.info("Flushing metrics data...")
             self.metrics_store.flush()
             logger.info("Flushing done")
+        # may happen if one of the load generators has detected that the user has cancelled the benchmark.
         elif isinstance(result, driver.BenchmarkCancelled):
             logger.info("User has cancelled the benchmark.")
             return False
@@ -170,7 +177,10 @@ class Benchmark:
             logger.info("Stopping engine has failed. Reason [%s]." % result.message)
             raise exceptions.RallyError(result.message, result.cause)
         else:
-            raise exceptions.RallyError("Mechanic has not stopped engine but instead [%s]. Terminating race without result." % str(result))
+            if self.ignore_unknown_return:
+                console.warn("Mechanic has not stopped engine but instead [%s]. Ignoring." % str(result), logger=logger)
+            else:
+                raise exceptions.RallyError("Mechanic has not stopped engine but instead [%s]. Terminating race without result." % str(result))
 
         self.metrics_store.flush()
         if not cancelled and not error: