+++ esrally/track/loader.py
@@ -49,8 +49,12 @@ def tracks(cfg):
 def list_tracks(cfg):
     console.println("Available tracks:\n")
     console.println(tabulate.tabulate(
-        tabular_data=[[t.name, t.short_description, t.default_challenge, ",".join(map(str, t.challenges))] for t in tracks(cfg)],
-        headers=["Name", "Description", "Default Challenge", "All Challenges"]))
+        tabular_data=[
+            [t.name, t.short_description, t.number_of_documents, convert.bytes_to_human_string(t.compressed_size_in_bytes),
+             convert.bytes_to_human_string(t.uncompressed_size_in_bytes), t.default_challenge,
+             ",".join(map(str, t.challenges))] for t in tracks(cfg)
+        ],
+        headers=["Name", "Description", "Documents", "Compressed Size", "Uncompressed Size", "Default Challenge", "All Challenges"]))
 
 
 def load_track(cfg):